--- packages/wallet/src/redux/protocols/helpers.ts
@@ -1,32 +1,0 @@
-import * as selectors from '../selectors';
-import { SharedData } from '../state';
-import { CommitmentType } from 'magmo-wallet-client/node_modules/fmg-core';
-
-export const channelIsClosed = (channelId: string, sharedData: SharedData): boolean => {
-  return (
-    channelHasConclusionProof(channelId, sharedData) ||
-    channelFinalizedOnChain(channelId, sharedData)
-  );
-};
-
-export const channelHasConclusionProof = (channelId: string, sharedData: SharedData): boolean => {
-  const channelState = selectors.getOpenedChannelState(sharedData, channelId);
-  const { lastCommitment, penultimateCommitment } = channelState;
-  return (
-    lastCommitment.commitment.commitmentType === CommitmentType.Conclude &&
-    penultimateCommitment.commitment.commitmentType === CommitmentType.Conclude
-  );
-};
-
-export const channelFinalizedOnChain = (channelId: string, sharedData: SharedData): boolean => {
-  const channelState = selectors.getAdjudicatorChannelState(sharedData, channelId);
-  return channelState && channelState.finalized;
-};
-
-export const isChannelDirectlyFunded = (channelId: string, sharedData: SharedData): boolean => {
-  const channelFundingState = selectors.getChannelFundingState(sharedData, channelId);
-  if (!channelFundingState) {
-    throw new Error(`No funding state for ${channelId}. Cannot determine funding type.`);
-  }
-  return channelFundingState.directlyFunded;
-};
