+++ packages/wallet/src/redux/protocols/reducer-helpers.ts
@@ -7,6 +7,9 @@ import { accumulateSideEffects } from '../outbox';
 import { SideEffects } from '../outbox/state';
 import { SharedData } from '../state';
 import { WalletProtocol } from '../types';
+import * as selectors from '../selectors';
+import { PlayerIndex } from '../types';
+import { CommitmentType } from 'fmg-core/lib/commitment';
 
 export const updateChannelState = (
   sharedData: SharedData,
@@ -62,3 +65,37 @@ export function theirAddress(channelState: channelStates.OpenedState) {
   const theirIndex = (channelState.ourIndex + 1) % channelState.participants.length;
   return channelState.participants[theirIndex];
 }
+
+export const channelIsClosed = (channelId: string, sharedData: SharedData): boolean => {
+  return (
+    channelHasConclusionProof(channelId, sharedData) ||
+    channelFinalizedOnChain(channelId, sharedData)
+  );
+};
+
+export const channelHasConclusionProof = (channelId: string, sharedData: SharedData): boolean => {
+  const channelState = selectors.getOpenedChannelState(sharedData, channelId);
+  const { lastCommitment, penultimateCommitment } = channelState;
+  return (
+    lastCommitment.commitment.commitmentType === CommitmentType.Conclude &&
+    penultimateCommitment.commitment.commitmentType === CommitmentType.Conclude
+  );
+};
+
+export const channelFinalizedOnChain = (channelId: string, sharedData: SharedData): boolean => {
+  const channelState = selectors.getAdjudicatorChannelState(sharedData, channelId);
+  return channelState && channelState.finalized;
+};
+
+export const isChannelDirectlyFunded = (channelId: string, sharedData: SharedData): boolean => {
+  const channelFundingState = selectors.getChannelFundingState(sharedData, channelId);
+  if (!channelFundingState) {
+    throw new Error(`No funding state for ${channelId}. Cannot determine funding type.`);
+  }
+  return channelFundingState.directlyFunded;
+};
+
+export const isFirstPlayer = (channelId: string, sharedData: SharedData) => {
+  const channelState = selectors.getChannelState(sharedData, channelId);
+  return channelState.ourIndex === PlayerIndex.A;
+};
